Add major functionality
At this point, we added a lot of the superficial features required.
We still need to change the image icons in the video entry list to
reflect the video actually loaded in. Bugs are that the video player
is temporarily not working and that the screen keeps refreshing on
its own.